backend/beanbuddy_ai/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "beanbuddy_ai"
version = "0.1.0"
dependencies = [
  "nvidia-nat[langchain]>=1.3.0a20250904",
]
requires-python = ">=3.11,<3.14"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'nat.components']
beanbuddy_ai = "beanbuddy_ai.register"